+++ mdbuild/glossary.py
@@ -2,6 +2,7 @@
 
 from __future__ import absolute_import
 
+import html
 import logging
 from operator import itemgetter
 import re
@@ -55,7 +56,7 @@ def _expand_term(term, key, pattern):
     try:
         return pattern % glossary['terms'][term][key]
     except KeyError:
-        logger.error("can't find '%s' for glossary entry '%s'" (key, term))
+        logger.error("can't find '%s' for glossary entry '%s'" % (key, term))
         raise
 
 
@@ -161,9 +162,9 @@ def get_glossary_link_processor(style):
         return GlossaryLinkMagic.replace_glossary_references
 
 
-class GlossaryLinkProcessor(object):
+class GlossaryLinkRenderer(object):
     """
-    Markdown processor for finding and rendering glossary links.
+    Filter for the renderer that finds and renders glossary links.
 
     Each subclass renders on specific target format
     """
@@ -184,14 +185,14 @@ class GlossaryLinkProcessor(object):
     @classmethod
     def additional_item_processing(cls, data):
         """Override for additional processing for each glossary item."""
-        pass
+        return data
 
     @classmethod
     def replace_callback(cls, match):
         """Replace each match of the regex."""
         data = cls.get_item_data(match)
         # do some additional stuff beyond replacing the glossary link inline:
-        cls.additional_item_processing(data)
+        data = cls.additional_item_processing(data)
         return cls.INLINE_TEMPLATE % data
 
     @classmethod
@@ -207,7 +208,7 @@ class GlossaryLinkProcessor(object):
         pass
 
 
-class GlossaryLinkMagic(GlossaryLinkProcessor):
+class GlossaryLinkMagic(GlossaryLinkRenderer):
     """
     Use the config parameter glossary_link_template
     as a template for replacing glossary links.
@@ -217,21 +218,27 @@ class GlossaryLinkMagic(GlossaryLinkProcessor):
     INLINE_TEMPLATE = ''
 
 
-class GlossaryLinkPlain(GlossaryLinkProcessor):
+class GlossaryLinkPlain(GlossaryLinkRenderer):
     """Remove all glossary links (replace with link title)."""
     INLINE_TEMPLATE = """%(title)s"""
 
 
-class GlossaryLinkTooltip(GlossaryLinkProcessor):
+class GlossaryLinkTooltip(GlossaryLinkRenderer):
     INLINE_TEMPLATE = """<dfn data-info="%(name)s: %(description)s">%(title)s</dfn>"""
 
+    @classmethod
+    def additional_item_processing(cls, item_data):
+        # buffer the explanation
+        item_data['description'] = html.escape(item_data['description'])
+        return item_data
 
-class GlossaryLinkUnderline(GlossaryLinkProcessor):
+
+class GlossaryLinkUnderline(GlossaryLinkRenderer):
     """Underline all glossary links."""
     INLINE_TEMPLATE = """`\\underline{%(title)s}`{=latex}"""
 
 
-class GlossaryLinkFootnote(GlossaryLinkProcessor):
+class GlossaryLinkFootnote(GlossaryLinkRenderer):
 
     INLINE_TEMPLATE = """%(title)s[^%(term)s]"""
     FOOTNOTE_TEXT_TEMPLATE = """[^%(term)s]: %(name)s: %(description)s"""
@@ -241,6 +248,7 @@ class GlossaryLinkFootnote(GlossaryLinkProcessor):
     def additional_item_processing(cls, item_data):
         # buffer the explanation
         cls.buffer[item_data['term']] = cls.FOOTNOTE_TEXT_TEMPLATE % item_data
+        return item_data
 
     @classmethod
     def glossary_post_processing(cls, target):