+++ mdbuild/macros/core.py
@@ -81,7 +81,7 @@ def process_macro(match, ignore_unknown=False):
 
 
 class MacroFilter(object):
-    """A markdown processor for expanding macros."""
+    """A renderer filter for expanding macros."""
     MACRO_PATTERN = re.compile(r'\{\{.*?\}\}')
 
     @classmethod