+++ docs-src/en/src/appendix/changelog.md
@@ -6,7 +6,8 @@
 -   new config-variables:
     -   read-next-navigation (true | false (default)): add navigation below content pages
     -   read-next-shortcuts ((true | false (default)))
--   cleaner code   
+-   cleaner code
+-   support markup for including or excluding content dependent on format, preset or cfg.edition
 
 ### v2.0.0 (2020-02-03)